requirements.txt
matplotlib==3.8.2
numpy==1.26.4
pandas==2.2.2
PyYAML==6.0.1
scikit_learn==1.3.2
scipy==1.13.0
torch==2.1.1
wandb==0.16.0
